Fork `hello-world` into `hello-constructors`
We're working to have a different, simpler "Hello, World" example. The
current "Hello, World" will henceforth be known as "Hello,
Constructors", since it will serve to introduce the concept of
constructors.

Everything in `hello-constructors/` is a copy of `hello-world/` with
only the replacement of "world" mentions with "constructor" mentions.

Since `hello-world` will soon be removed from this repo and factored out
into a separate repo, this PR also changes all the tests to use the new
`hello-constructors` instead of `hello-world`.