Provide a Dev Container (#9)
* Provide a Dev Container

Before this change, users had to provide their own system to work with
our Resemble examples. This change introduces a Dev Container that
provides an official development environment that is built especially to
work for Resemble.

Details for usage have been added to the `README.md`.

For the tooling that builds the Dev Container's image, see
https://github.com/reboot-dev/resemble-devcontainer/pull/2.

* Addressing review comments